Does this message make you curious?
readme is the default branch for asasvirtuais/asasvirtuais:
My multi trunk repository, which is also my github homepage repository.
Since, GitHub uses the default branch as the branch for the homepage PIN
I made the main branch an "orphan" branch :)

Check the seed branch to see the starting point of my multi trunk
repository. Which is a template branch that ensures all my projects has
the bare minimum of my personal codebase.